Add a launcher wrapper for the guarded publish path
The publishing doc already names this command, but it had shipped as
only a raw TypeScript entry point without the wrapper the other
launchers have.

An extensionless wrapper would shadow the entry point for the test that
imports its checks, so move that tested surface into a shared library
both the entry point and its test import.
